open_exchange_rates v0.2.0 OpenExchangeRates

This module contains all the helper methods for converting currencies

Functions

available_currencies()

Specs

available_currencies :: [String.t]

Returns a list of all available currencies.

example

iex> OpenExchangeRates.available_currencies |> Enum.take(10)
["AWG", "NAD", "INR", "LAK", "BOB", "MOP", "QAR", "SDG", "TMT", "BRL"]

conversion_rate(from, to)

Specs

conversion_rate(String.t | Atom.t, String.t | Atom.t) ::
  {:ok, Float.t} |
  {:error, String.t}

Get the conversion rate for a between two currencies”

Example

iex> OpenExchangeRates.conversion_rate(:EUR, :GBP)
{:ok, 0.8481186252771619}

convert(value, from, to)

Specs

convert(Float.t, String.t | Atom.t, String.t | Atom.t) ::
  {:ok, Float.t} |
  {:error, String.t}

Will convert a price from once currency to another

example

iex> OpenExchangeRates.convert(100.00, :EUR, :GBP)
{:ok, 84.81186252771619}

convert_cents(value, from, to)

Specs

convert_cents(Integer.t, String.t | Atom.t, String.t | Atom.t) ::
  {:ok, Integer.t} |
  {:error, String.t}

Will convert cents from once currency to another

example

iex> OpenExchangeRates.convert_cents(100, :GBP, :AUD)
{:ok, 172}